+++ Reqnroll.VisualStudio/Editor/Commands/ReqnrollVsCommands.cs
@@ -8,5 +8,6 @@ public static class ReqnrollVsCommands
     public const int FindStepDefinitionUsagesCommandId = 0x0101;
     public const int RenameStepCommandId = 0x0103;
     public const int GoToHookCommandId = 0x0104;
+    public const int FindUnusedStepDefinitionsCommandId = 0x0102;
     public static readonly Guid DefaultCommandSet = new("7fd3ed5d-2cf1-4200-b28b-cf1cc6b00c5a");
 }